Move HasCapabilityNoAlert to a synced file so AgentService satisfies the cloud interfaces
The synced AgentServicer / AgentServicerForVerification handler interfaces require
HasCapabilityNoAlert, but the honeytoken feature defined it on AgentService in
agent_service.go, which the AIM Cloud sync treats as protected (keeps a divergent
copy). The cloud build therefore failed across every handler with '*AgentService
does not implement AgentServicer (missing method HasCapabilityNoAlert)'. Relocating
the verbatim method to a synced file lets the cloud's AgentService satisfy the
interfaces. Its body only uses capabilityRepo and matchesCapability, both present
since the initial release. No logic change; identical public build. Follow-up to #319.
